Pressing the button will change the field to the next player
When pressing on a field, the image is set to the next player sprite and background color of the field is set to the color of next player specified.
Currently, there is no check for win state and neither for if the field is already filled in by a different cat, resulting in a state, where the field can be overwritten infinitelly many times.